+++ src/imgaug/cl/kernel/contrast_stretch.cl
@@ -11,14 +11,14 @@ __kernel void contrast_stretch(  __global unsigned char *a,
                                const unsigned short channel
 )
 {
-    int id_x = get_global_id(0);
-    int id_y = get_global_id(1);
-    int id_z = get_global_id(2);
+    unsigned short id_x = get_global_id(0);
+    unsigned short id_y = get_global_id(1);
+    unsigned short id_z = get_global_id(2);
     if (id_x >= width || id_y >= height || id_z >= channel) return;
 
-    int pixIdx = id_x + id_y* width + id_z * width * height;
+    unsigned short pixIdx = id_x + id_y* width + id_z * width * height;
 
-    int res = a[pixIdx] - min) * (new_max - new_min)/((max - min) * 1.0) + new_min ;
+    unsigned short res = a[pixIdx] - min) * (new_max - new_min)/((max - min) * 1.0) + new_min ;
 
     c[pixIdx] = saturate_8u(res);
 }
